move shell to shells.shell to avoid unwanted init
