+++ pyproject.toml
@@ -27,8 +27,8 @@ readme = {file = ["README.rst"]}
 packages = [
     "xonsh",
     "xonsh.ply.ply",
-    "xonsh.shell",
-    "xonsh.shell.ptk_shell",
+    "xonsh.shells",
+    "xonsh.shells.ptk_shell",
     "xonsh.procs",
     "xonsh.platform",
     "xonsh.parsers",
@@ -243,11 +243,11 @@ convention = "numpy"
 "xonsh/style_tools.py" = ["F821"]
 "xonsh/xoreutils/*.py" = ["E722"]
 "xonsh/completers/python.py" = ["E722"]
-"xonsh/shell/ptk_shell/__init__.py" = ["E731"]
+"xonsh/shells/ptk_shell/__init__.py" = ["E731"]
 "xonsh/ast.py" = ["F401"]
-"xonsh/shell/readline_shell.py" = ["F401"]
+"xonsh/shells/readline_shell.py" = ["F401"]
 "xonsh/commands_cache.py" = ["F841"]
-"xonsh/shell/ptk_shell/key_bindings.py" = ["F841"]
+"xonsh/shells/ptk_shell/key_bindings.py" = ["F841"]
 "xonsh/tools.py" = [
     "E731",
 ]
