+++ xonsh/shells/__init__.py
@@ -1,0 +1,1 @@
+"""The xonsh interactive shells"""
